third_party/blink/web_tests/fast/js/pic/cached-getter-dictionary-and-proto-expected.txt

This page tests cached access to properties of dictionary objects and objects with changing prototypes. If the test passes, you'll see a series of PASS messages below.

PASS: getX(o) should be 0 and is.

PASS: getProtoX(o) should be 2 and is.

PASS: getProtoX(o) should be undefined and is.
